fix(ZH-81): 로그인/로그아웃 버튼 분리
